Add inline form errors
Instead of rendering form errors as a list at the top, I used a custom
form builder to structure the form fields with helper text of the error
message, and ensured the HTML matched the visual error states of
pico.css.

Now all I have to do is use form_with_inline_errors in the future, and
error messages will appear beneath the fields as necessary!

The one drawback with this approach is I need to remember to add any
form helpers I use to the form builder, but that's an okay tradeoff for
now.